db persistence

+++ Backend/src/main/java/com/novabank/repository/PaymentHistoryRepository.java
@@ -12,5 +12,5 @@ import java.util.List;
 public interface PaymentHistoryRepository extends CrudRepository<PaymentHistory,Integer> {
 
     @Query(value = "SELECT * FROM v_payments WHERE user_id = :user_id ORDER BY payment_id DESC",nativeQuery = true)
-    List<PaymentHistory> getPaymentsRecordsById(@Param("user_id")int user_id);
+    List<PaymentHistory> getPaymentsRecordsById(@Param("user_id")Long user_id);
 }
